Refactored release/0.1.0 by extracting df funct
Almost all Dialogflow functionality that was within index.js has been
moved to DialogflowController, which acts as our interface to the
Dialogflow agent. The DialogflowController 'class' takes an agentId and
sessionId (meaning potentially multiple agents and sessions can be
established simulatenously) and each establish their own client and
client session. So far, DialogflowController only allows for querying,
specifically, via queryAgent(queryString, callback). The queryString is
sent to the agent and the callback is ran on the response when it
eventually arrives. All Dialogflow specific details were not extracted
from index.js (i.e., the callback function 'printResponse' still knows
the properties of the agent's response object). This was left as is
becaause replacing these parts of code requires more methods on the
DialogflowController prototype that can't be easily accessed (from my
current knowledge).