Added basic twitch integration
Using the tmi.js package, the base of twitch integration was added. Any
messages posted in the twitch chat are read by the application and are
automatically sent to the Dialogflow agent. Both the chat message and
the agent's response are printed to console. Note that there does not
seem to be any imposed delay between separate chat messages, which can
both good and bad. This is good in that all messages actually can be
processed quick enough for the agent listen to them all! Bad thing is
that this is potentially dangerous in that a fast chat could overload
either the application or the quota imposed by the Dialogflow team.
Something to keep an eye on.