Created TwitchBot.js, basic read-resp. funct. impl
Using the tmi.js library, created a very simple TwitchBot 'class' that
reads every chat message, sends the message (depending on conditions) to
dialogflow and writes the agent's response to the chat. Things to note:
1. Twitch bot requires a reference to the DialogflowController that
index.js declares & constructs. 2. TwitchBot only contains the 1 public
method 'start', which sets up the 'onMessage' event in a very broken
down way (for readability sake). 4 inner functions are created and are
composed to complete the task well. However, the readibility starts to
decrease with all of the calls and binds using 'this', which are
necessary for the inner functions.