Decoupled DFController and TwitchBot
Previously, TwitchBot had a very strong relationship with the
DialogflowController class. Specifically, it held an instance of this
class and used methods directly from it. While for this specific
project, this coupling doesn't quite matter since it will always be
Dialogflow, it was still decided that the two were to be decoupled.
Specifically, it was the thought that there should be the added
flexibility of being able to choose the service that will process the
twitch messages, maybe in a future and separate project. So, instead of
providing an instance of DialogflowController, you would provide the
callback methods specified in index.js.